+++ Firestore/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - [feature] Added support for Pipeline expressions `arraySlice`, `arrayFilter`, `arrayTransform` and `arrayTransformWithIndex`. (#16001)
+- [feature] Added support for the `forceIndex` option on Pipeline `collection` and `collectionGroup` stages.
 
 # 12.12.0
 - [feature] Added support for the `parent` Pipeline expression. (#16010)